backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["kein <fardz9@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
